+++ com.unity.shadergraph/Editor/Data/Nodes/Artistic/Normal/NormalDeriveZNode.cs
@@ -3,25 +3,25 @@ using UnityEngine;
 
 namespace UnityEditor.ShaderGraph
 {
-    [Title("Artistic", "Normal", "Normal Derive Z")]
-    public class NormalDeriveZNode : CodeFunctionNode
+    [Title("Artistic", "Normal", "Normal Reconstruct Z")]
+    public class NormalReconstructZNode : CodeFunctionNode
     {
-        public NormalDeriveZNode()
+        public NormalReconstructZNode()
         {
-            name = "Normal Derive Z";
+            name = "Normal Reconstruct Z";
         }
 
         public override string documentationURL
         {
-            get { return "https://github.com/Unity-Technologies/ShaderGraph/wiki/Normal-Derive-Z-Node"; }
+            get { return "https://github.com/Unity-Technologies/ShaderGraph/wiki/Normal-Reconstruct-Z-Node"; }
         }
 
         protected override MethodInfo GetFunctionToConvert()
         {
-            return GetType().GetMethod("NormalDeriveZ", BindingFlags.Static | BindingFlags.NonPublic);
+            return GetType().GetMethod("NormalReconstructZ", BindingFlags.Static | BindingFlags.NonPublic);
         }
 
-        static string NormalDeriveZ(
+        static string NormalReconstructZ(
             [Slot(0, Binding.None)] Vector2 In,
             [Slot(2, Binding.None, ShaderStageCapability.Fragment)] out Vector3 Out)
             
@@ -30,8 +30,8 @@ namespace UnityEditor.ShaderGraph
             return
                 @"
 {
-    float deriveZ = sqrt(1 - ( In.x * In.x + In.y * In.y));
-    float3 normalVector = float3(In.x, In.y, deriveZ);
+    float reconstructZ = sqrt(1 - ( In.x * In.x + In.y * In.y));
+    float3 normalVector = float3(In.x, In.y, reconstructZ);
     Out = normalize(normalVector);
 }";
         }